+++ docs/CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+### Added
+- Add new flaw reference type "UPSTREAM"
+
 ## [4.4.1] - 2024-10-17
 ### Added
 - Auto-reset CVSS validation flag on NVD CVSS removal (OSIDB-3407)